IS-3295-3: Add avslag fom date to innstilling om avslag uten forhandsvarsel
