+++ src/routers/AppRouter.tsx
@@ -46,8 +46,8 @@ import ArbeidsuforhetAvslag from "@/sider/arbeidsuforhet/avslag/ArbeidsuforhetAv
 import ArbeidsuforhetIkkeAktuell from "@/sider/arbeidsuforhet/ikkeaktuell/ArbeidsuforhetIkkeAktuell";
 import SendForhandsvarselSkjema from "@/sider/arbeidsuforhet/SendForhandsvarselSkjema";
 import ArbeidsuforhetSide from "@/sider/arbeidsuforhet/ArbeidsuforhetSide";
-import InnstillingUtenForhandsvarsel from "@/sider/arbeidsuforhet/InnstillingUtenForhandsvarsel";
 import { DinSykmeldingSide } from "@/sider/sykmeldinger/container/DinSykmeldingSide";
+import InnstillingUtenForhandsvarsel from "@/sider/arbeidsuforhet/innstillingutenforhandsvarsel/InnstillingUtenForhandsvarsel";
 
 export const appRoutePath = "/sykefravaer";
